Move into keypair package
