+++ keypair/generator.go
@@ -1,6 +1,4 @@
-package keypairgen
-
-import "github.com/stellar/go/keypair"
+package keypair
 
 // Generator generates new keys with the underlying source. The underlying
 // source defaults to the RandomSource if not specified.
@@ -16,20 +14,20 @@ func (g *Generator) getSource() Source {
 }
 
 // Generate returns a new key using the underlying source.
-func (g *Generator) Generate() (*keypair.Full, error) {
+func (g *Generator) Generate() (*Full, error) {
 	return g.getSource().Generate()
 }
 
 // Source provides keys.
 type Source interface {
-	Generate() (*keypair.Full, error)
+	Generate() (*Full, error)
 }
 
 // RandomSource provides new keys that are randomly generated using the
-// keypair.Random function.
+// Random function.
 type RandomSource struct{}
 
 // Generated returns a new key using keypair.Random.
-func (RandomSource) Generate() (*keypair.Full, error) {
-	return keypair.Random()
+func (RandomSource) Generate() (*Full, error) {
+	return Random()
 }
