+++ keypair/generator_test.go
@@ -1,17 +1,16 @@
-package keypairgen_test
+package keypair_test
 
 import (
 	"testing"
 
 	"github.com/stellar/go/keypair"
-	"github.com/stellar/go/keypairgen"
-	"github.com/stellar/go/keypairgen/keypairgentest"
+	"github.com/stellar/go/keypair/keypairtest"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestGenerator_Generate_sourceNotSet(t *testing.T) {
-	g := keypairgen.Generator{}
+	g := keypair.Generator{}
 	k1, err := g.Generate()
 	require.NoError(t, err)
 	t.Log("k1", k1.Address(), k1.Seed())
@@ -22,7 +21,7 @@ func TestGenerator_Generate_sourceNotSet(t *testing.T) {
 }
 
 func TestGenerator_Generate_sourceNotSetPtrNil(t *testing.T) {
-	g := (*keypairgen.Generator)(nil)
+	g := (*keypair.Generator)(nil)
 	k1, err := g.Generate()
 	require.NoError(t, err)
 	t.Log("k1", k1.Address(), k1.Seed())
@@ -33,12 +32,12 @@ func TestGenerator_Generate_sourceNotSetPtrNil(t *testing.T) {
 }
 
 func TestGenerator_Generate_sourceSet(t *testing.T) {
-	s := keypairgentest.SliceSource{
+	s := keypairtest.SliceSource{
 		keypair.MustRandom(),
 		keypair.MustRandom(),
 		keypair.MustRandom(),
 	}
-	g := keypairgen.Generator{
+	g := keypair.Generator{
 		Source: &s,
 	}
 	k1, err := g.Generate()
